git-svn-id: http://svn.code.sf.net/p/shopizer/code/shopizer-portlet@2748 e80ee2c7-76f3-4005-aeaf-98396743f455

+++ sm-shop/src/main/java/com/salesmanager/web/admin/controller/categories/CategoryController.java
@@ -334,7 +334,7 @@ public class CategoryController {
 			
 			Category category = categoryService.getById(id);
 			
-			if(category==null || category.getMerchantStore().getId()!=store.getId()) {
+			if(category==null || category.getMerchantStore().getId().intValue() !=store.getId().intValue() ) {
 
 				resp.setStatusMessage(messages.getMessage("message.unauthorized", locale));
 				resp.setStatus(AjaxResponse.RESPONSE_STATUS_FAIURE);
